requirement.txt
django-cors-headers==4.2.0
django-filter==23.3
drf-yasg==1.21.7
Pillow==10.0.1
pip==22.0.2
pipdeptree==2.13.0
setuptools==59.6.0
whitenoise==6.5.0
